Add drawCigar setting
